refactor(utils): hoist `utils::utils` into `utils`
