+++ src/cli/common.rs
@@ -27,7 +27,7 @@ use crate::{
     notifications::Notification,
     process::{terminalsource, Process},
     toolchain::{DistributableToolchain, LocalToolchainName, Toolchain, ToolchainName},
-    utils::{notifications as util_notifications, notify::NotificationLevel, utils},
+    utils::{self, notifications as util_notifications, notify::NotificationLevel},
 };
 
 pub(crate) const WARN_COMPLETE_PROFILE: &str = "downloading with complete profile isn't recommended unless you are a developer of the rust language";
